Redo structure to es6 modules
