+++ index.html
@@ -112,7 +112,7 @@
         </div>
 
     </div>
-    <script type="module" src="js/feed.js"></script>
+    <script type="module" src="js/index.mjs"></script>
 </body>
 
 </html>
