+++ login.html
@@ -75,7 +75,7 @@
                     placeholder="Password">
             </div>
             <!-- Errors -->
-            <div id="errors-login" class="hidden text-red-500 mt-3 font-bold">Errors here</div>
+            <div id="errors-login" class="hidden text-red-500 mt-3 font-bold"></div>
             <!-- Submit button -->
             <button
                 class="bg-orange-300 py-3 px-6 block w-full rounded-md text-white font-bold mt-3 hover:bg-orange-500 transition-all duration-200"
@@ -86,7 +86,7 @@
                 class="text-neutral-500 inline-block mt-3 font-medium hover:text-orange-500">Register here</a>
         </div>
     </div>
-    <script type="module" src="js/login.js"></script>
+    <script type="module" src="js/index.mjs"></script>
 </body>
 
 </html>
