+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "npx tailwindcss build -i ./src/css/styles.css -o ./css/styles.css",
-    "watch": "npx tailwindcss build -i ./src/css/styles.css -o ./css/styles.css --watch"
+    "watch": "npx tailwindcss build -i ./src/css/styles.css -o ./css/styles.css --watch & live-server"
   },
   "repository": {
     "type": "git",
