+++ profile.html
@@ -105,7 +105,7 @@
             </div>
         </div>
     </div>
-    <script type="module" src="js/profile.js"></script>
+    <script type="module" src="js/index.mjs"></script>
 </body>
 
 </html>
