+++ register.html
@@ -119,7 +119,7 @@
                 have an account?</a>
         </div>
     </div>
-    <script type="module" src="js/registration.js"></script>
+    <script type="module" src="js/index.mjs"></script>
 </body>
 
 </html>
