+++ single.html
@@ -108,7 +108,7 @@
             <div id="posts-container" class="mb-8">
             </div>
         </div>
-        <script type="module" src="js/single.js"></script>
+        <script type="module" src="js/index.mjs"></script>
 </body>
 
 </html>
